+++ Modules/Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation/Templates/DtoValidator/DtoValidatorTemplatePartial.cs
@@ -7,9 +7,14 @@ using Intent.Modules.Blazor.HttpClients.Templates.DtoContract;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.Templates;
 using Intent.Modules.Contracts.Clients.Shared;
 using Intent.Modules.FluentValidation.Shared;
 using Intent.RoslynWeaver.Attributes;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Editing;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: DefaultIntentManaged(Mode.Ignore, Targets = Targets.Usings)]
@@ -18,8 +23,7 @@ using Intent.RoslynWeaver.Attributes;
 namespace Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation.Templates.DtoValidator
 {
     [IntentManaged(Mode.Merge)]
-    public partial class DtoValidatorTemplate : CSharpTemplateBase<DTOModel>, ICSharpFileBuilderTemplate
-, IFluentValidationTemplate
+    public partial class DtoValidatorTemplate : CSharpTemplateBase<DTOModel>, ICSharpFileBuilderTemplate, IFluentValidationTemplate
     {
         public const string TemplateId = "Intent.Blazor.HttpClients.Dtos.FluentValidation.DtoValidator";
 
@@ -56,6 +60,47 @@ namespace Intent.Modules.Blazor.HttpClients.Dtos.FluentValidation.Templates.DtoV
         {
             return CSharpFile.ToString();
         }
+    
+        public override RoslynMergeConfig ConfigureRoslynMerger()
+        {
+            return new RoslynMergeConfig(new TemplateMetadata(Id, "2.0"), new ConstructorSignatureMigration());
+        }
 
+        public class ConstructorSignatureMigration : ITemplateMigration
+        {
+            public string Execute(string currentText)
+            {
+                var syntaxTree = CSharpSyntaxTree.ParseText(currentText);
+
+                var root = syntaxTree.GetRoot();
+                using var workspace = new AdhocWorkspace();
+            
+                var services = workspace.Services;
+                var editor = new SyntaxEditor(root, services);
+            
+                var attributeLists = root.DescendantNodes()
+                    .OfType<ConstructorDeclarationSyntax>()
+                    .SelectMany(c => c.AttributeLists);
+            
+                foreach (var attributeList in attributeLists)
+                {
+                    var intentManagedAttributes = attributeList.Attributes
+                        .Where(attr => attr.Name.ToString().EndsWith("IntentManaged"));
+
+                    foreach (var attribute in intentManagedAttributes)
+                    {
+                        var newArgumentList = SyntaxFactory.ParseAttributeArgumentList("(Mode.Merge)");
+
+                        editor.ReplaceNode(attribute, attribute.WithArgumentList(newArgumentList));
+                    }
+                }
+
+                var newRoot = editor.GetChangedRoot();
+
+                return newRoot.ToFullString();
+            }
+
+            public TemplateMigrationCriteria Criteria => TemplateMigrationCriteria.Upgrade(1, 2);
+        }
     }
 }